pyproject.toml
[project]
name = "fast-data-loader"
version = "0.1.0"
description = "A data loader that quickly loads files to a set postgres database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.5",
    "duckdb>=1.0.0",
    "python-dotenv>=1.0.1",
    "psycopg>=3.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fast_data_loader"]
